기능 : 서비스 단위 테스트 각 api 에 추가.

+++ EarnPointAPI/src/main/java/com/kakaopay/earnpointapi/api/dto/EarnRequestDTO.java
@@ -17,7 +17,7 @@ public class EarnRequestDTO {
     private String storeName;
 
     @ApiModelProperty(value = "요청 카테고리", required = true)
-    private Integer categoryId;
+    private Long categoryId;
 
     @ApiModelProperty(value = "요청 바코드", required = true)
     private String barcode;
@@ -27,7 +27,7 @@ public class EarnRequestDTO {
 
     public EarnRequestDTO() { }
 
-    public EarnRequestDTO(String accessType, Integer categoryId, String barcode, Integer earnPoint) {
+    public EarnRequestDTO(String accessType, Long categoryId, String barcode, Integer earnPoint) {
         this.accessType = accessType;
         this.categoryId = categoryId;
         this.barcode = barcode;
