+++ UsePointAPI/src/main/java/com/kakaopay/usepointapi/api/dto/UseRequestDTO.java
@@ -17,7 +17,7 @@ public class UseRequestDTO {
     private String storeName;
 
     @ApiModelProperty(value = "요청 카테고리", required = true)
-    private Integer categoryId;
+    private Long categoryId;
 
     @ApiModelProperty(value = "요청 바코드", required = true)
     private String barcode;
@@ -27,7 +27,7 @@ public class UseRequestDTO {
 
     public UseRequestDTO() { }
 
-    public UseRequestDTO(String accessType, Integer categoryId, String barcode, Integer usePoint) {
+    public UseRequestDTO(String accessType, Long categoryId, String barcode, Integer usePoint) {
         this.accessType = accessType;
         this.categoryId = categoryId;
         this.barcode = barcode;
