+++ UsePointAPI/src/main/java/com/kakaopay/usepointapi/api/vo/PointVo.java
@@ -19,7 +19,7 @@ public class PointVo {
     private String barcode;
 
     @JoinColumn
-    private Integer category;
+    private Long category;
 
     @JoinColumn
     @Column(name = "store_id")
@@ -32,7 +32,7 @@ public class PointVo {
     private LocalDateTime createdAt;
 
 
-    public void update(String barcode, Integer category, Integer point) {
+    public void update(String barcode, Long category, Integer point) {
         this.barcode = barcode;
         this.category = category;
         this.point = point;
